langchain_RAG/requirement.txt
openai
langchain
langchain-community
pypdf
yt_dlp
pydub
python-dotenv
beautifulsoup4